+++ src/crossterm.rs
@@ -22,12 +22,12 @@ impl Crossterm {
         crate::style::TerminalColor::new()
     }
 
-    /// Creates a new `StyledObject`.
+    /// Creates a new `StyledContent`.
     #[cfg(feature = "style")]
-    pub fn style<D>(&self, val: D) -> crate::style::StyledObject<D>
+    pub fn style<D>(&self, val: D) -> crate::style::StyledContent<D>
     where
         D: Display + Clone,
     {
-        crate::style::ObjectStyle::new().apply_to(val)
+        crate::style::ContentStyle::new().apply(val)
     }
 }