+++ src/style/enums/attribute.rs
@@ -93,7 +93,7 @@ pub enum Attribute {
     /// Turns off the text blinking (`SlowBlink` or `RapidBlink`).
     NoBlink = 25,
     /// Turns off the `Reverse` attribute.
-    NoInverse = 27, // TODO Shouldn't we rename this to `NoReverse`? Or `Reverse` to `Inverse`?
+    NoReverse = 27,
     /// Turns off the `Hidden` attribute.
     NoHidden = 28,
     /// Turns off the `CrossedOut` attribute.