+++ src/style/styledcontent.rs
@@ -1,13 +1,13 @@
-//! This module contains the logic to style an object that contains some 'content' which can be styled.
+//! This module contains the logic to style some content.
 
 use std::fmt::{self, Display, Formatter};
 use std::result;
 
 use crate::queue;
 
-use super::{Attribute, Color, Colorize, ObjectStyle, ResetColor, SetAttr, SetBg, SetFg, Styler};
+use super::{Attribute, Color, Colorize, ContentStyle, ResetColor, SetAttr, SetBg, SetFg, Styler};
 
-/// A styled object.
+/// A styled content.
 ///
 /// # Examples
 ///
@@ -17,54 +17,69 @@ use super::{Attribute, Color, Colorize, ObjectStyle, ResetColor, SetAttr, SetBg,
 /// let styled = style("Hello there")
 ///     .with(Color::Yellow)
 ///     .on(Color::Blue)
-///     .attr(Attribute::Bold);
+///     .attribute(Attribute::Bold);
 ///
 /// println!("{}", styled);
 /// ```
 #[derive(Clone)]
-pub struct StyledObject<D: Display + Clone> {
-    /// The object style (colors, content attributes).
-    pub object_style: ObjectStyle,
-    /// An object to apply the style on.
-    pub content: D,
+pub struct StyledContent<D: Display + Clone> {
+    /// The style (colors, content attributes).
+    style: ContentStyle,
+    /// A content to apply the style on.
+    content: D,
 }
 
-impl<'a, D: Display + 'a + Clone> StyledObject<D> {
+impl<'a, D: Display + 'a + Clone> StyledContent<D> {
+    /// Creates a new `StyledContent`.
+    pub fn new(style: ContentStyle, content: D) -> StyledContent<D> {
+        StyledContent { style, content }
+    }
+
     /// Sets the foreground color.
-    pub fn with(mut self, foreground_color: Color) -> StyledObject<D> {
-        self.object_style = self.object_style.fg(foreground_color);
+    pub fn with(mut self, foreground_color: Color) -> StyledContent<D> {
+        self.style = self.style.foreground(foreground_color);
         self
     }
 
     /// Sets the background color.
-    pub fn on(mut self, background_color: Color) -> StyledObject<D> {
-        self.object_style = self.object_style.bg(background_color);
+    pub fn on(mut self, background_color: Color) -> StyledContent<D> {
+        self.style = self.style.background(background_color);
         self
     }
 
     /// Adds the attribute.
     ///
     /// You can add more attributes by calling this method multiple times.
-    pub fn attr(mut self, attr: Attribute) -> StyledObject<D> {
-        self.object_style.add_attr(attr);
+    pub fn attribute(mut self, attr: Attribute) -> StyledContent<D> {
+        self.style = self.style.attribute(attr);
         self
+    }
+
+    /// Returns the content.
+    pub fn content(&self) -> &D {
+        &self.content
+    }
+
+    /// Returns the style.
+    pub fn style(&self) -> &ContentStyle {
+        &self.style
     }
 }
 
-impl<D: Display + Clone> Display for StyledObject<D> {
+impl<D: Display + Clone> Display for StyledContent<D> {
     fn fmt(&self, f: &mut Formatter<'_>) -> result::Result<(), fmt::Error> {
         let mut reset = false;
 
-        if let Some(bg) = self.object_style.bg_color {
+        if let Some(bg) = self.style.bg_color {
             queue!(f, SetBg(bg)).map_err(|_| fmt::Error)?;
             reset = true;
         }
-        if let Some(fg) = self.object_style.fg_color {
+        if let Some(fg) = self.style.fg_color {
             queue!(f, SetFg(fg)).map_err(|_| fmt::Error)?;
             reset = true;
         }
 
-        for attr in self.object_style.attrs.iter() {
+        for attr in self.style.attrs.iter() {
             queue!(f, SetAttr(*attr)).map_err(|_| fmt::Error)?;
             reset = true;
         }
@@ -79,7 +94,7 @@ impl<D: Display + Clone> Display for StyledObject<D> {
     }
 }
 
-impl<D: Display + Clone> Colorize<D> for StyledObject<D> {
+impl<D: Display + Clone> Colorize<D> for StyledContent<D> {
     // foreground colors
     def_color!(fg_color: black => Color::Black);
     def_color!(fg_color: dark_grey => Color::DarkGrey);
@@ -117,7 +132,7 @@ impl<D: Display + Clone> Colorize<D> for StyledObject<D> {
     def_color!(bg_color: on_grey => Color::Grey);
 }
 
-impl<D: Display + Clone> Styler<D> for StyledObject<D> {
+impl<D: Display + Clone> Styler<D> for StyledContent<D> {
     def_attr!(reset => Attribute::Reset);
     def_attr!(bold => Attribute::Bold);
     def_attr!(underlined => Attribute::Underlined);
@@ -133,24 +148,26 @@ impl<D: Display + Clone> Styler<D> for StyledObject<D> {
 
 #[cfg(test)]
 mod tests {
-    use super::{Attribute, Color, ObjectStyle};
+    use super::{Attribute, Color, ContentStyle};
 
     #[test]
     fn test_set_fg_bg_add_attr() {
-        let mut object_style = ObjectStyle::new().fg(Color::Blue).bg(Color::Red);
-        object_style.add_attr(Attribute::Reset);
+        let style = ContentStyle::new()
+            .foreground(Color::Blue)
+            .background(Color::Red)
+            .attribute(Attribute::Reset);
 
-        let mut styled_object = object_style.apply_to("test");
+        let mut styled_content = style.apply("test");
 
-        styled_object = styled_object
+        styled_content = styled_content
             .with(Color::Green)
             .on(Color::Magenta)
-            .attr(Attribute::NoItalic);
+            .attribute(Attribute::NoItalic);
 
-        assert_eq!(styled_object.object_style.fg_color, Some(Color::Green));
-        assert_eq!(styled_object.object_style.bg_color, Some(Color::Magenta));
-        assert_eq!(styled_object.object_style.attrs.len(), 2);
-        assert_eq!(styled_object.object_style.attrs[0], Attribute::Reset);
-        assert_eq!(styled_object.object_style.attrs[1], Attribute::NoItalic);
+        assert_eq!(styled_content.style.fg_color, Some(Color::Green));
+        assert_eq!(styled_content.style.bg_color, Some(Color::Magenta));
+        assert_eq!(styled_content.style.attrs.len(), 2);
+        assert_eq!(styled_content.style.attrs[0], Attribute::Reset);
+        assert_eq!(styled_content.style.attrs[1], Attribute::NoItalic);
     }
 }